build/configs/rtl8721csm_download.sh : Add condition for loadable apps
1) Add conditions for loadable apps about checking each binary exists
Without this condition, it makes error like below.
cp: cannot stat '/root/tizenrt/build/configs/../../build/output/bin/micom': No such file or directory
cp: cannot stat '/root/tizenrt/build/configs/../../build/output/bin/wifi': No such file or directory

2) If there is no user binary, make error.

+++ build/configs/rtl8721csm/rtl8721csm_download.sh
@@ -24,6 +24,7 @@ IMG_TOOL_PATH=${TOOL_PATH}/image_tool
 OS_PATH=${TOP_PATH}/os
 CONFIG=${OS_PATH}/.config
 source ${CONFIG}
+APP_NUM=0
 
 function pre_download()
 {
@@ -31,8 +32,21 @@ function pre_download()
 	cp -p ${BIN_PATH}/km0_boot_all.bin ${IMG_TOOL_PATH}/km0_boot_all.bin
 	cp -p ${BIN_PATH}/km4_boot_all.bin ${IMG_TOOL_PATH}/km4_boot_all.bin
 	cp -p ${BIN_PATH}/${KERNEL_BIN_NAME} ${IMG_TOOL_PATH}/${KERNEL_BIN_NAME}
-	cp -p ${BIN_PATH}/micom ${IMG_TOOL_PATH}/micom
-	cp -p ${BIN_PATH}/wifi ${IMG_TOOL_PATH}/wifi
+	if [ "${CONFIG_APP_BINARY_SEPARATION}" == "y" ]; then
+		if test -f "${BIN_PATH}/micom"; then
+			APP_NUM=$(($APP_NUM+1))
+			cp -p ${BIN_PATH}/micom ${IMG_TOOL_PATH}/micom
+		fi
+		if test -f "${BIN_PATH}/wifi"; then
+			APP_NUM=$(($APP_NUM+1))
+			cp -p ${BIN_PATH}/wifi ${IMG_TOOL_PATH}/wifi
+		fi
+		if [ ${APP_NUM} -eq 0 ]; then
+			echo "No User Binary."
+			post_download
+			exit 1
+		fi
+	fi
 	if test -f "${BIN_PATH}/${CONFIG_COMMON_BINARY_NAME}"; then
 		cp -p ${BIN_PATH}/${CONFIG_COMMON_BINARY_NAME} ${IMG_TOOL_PATH}/${CONFIG_COMMON_BINARY_NAME}
 	fi
@@ -62,8 +76,12 @@ function post_download()
 	[ -e ${BL1}.bin ] && rm ${BL1}.bin
 	[ -e ${BL2}.bin ] && rm ${BL2}.bin
 	[ -e ${KERNEL_BIN_NAME} ] && rm ${KERNEL_BIN_NAME}
-	[ -e micom ] && rm micom
-	[ -e wifi ] && rm wifi
+	if test -f "micom"; then
+		[ -e micom ] && rm micom
+	fi
+	if test -f "wifi"; then
+		[ -e wifi ] && rm wifi
+	fi
 	if test -f "${CONFIG_COMMON_BINARY_NAME}"; then
 		[ -e ${CONFIG_COMMON_BINARY_NAME} ] && rm ${CONFIG_COMMON_BINARY_NAME}
 	fi
